+++ src/constants.ts
@@ -1,7 +1,7 @@
 import { css } from 'lit';
 
 /** current version of the card. */
-export const CARD_VERSION = '1.0.25';
+export const CARD_VERSION = '1.0.26';
 
 /** SpotifyPlus integration domain identifier. */
 export const DOMAIN_SPOTIFYPLUS = 'spotifyplus';
@@ -48,6 +48,9 @@ export const PLAYER_CONTROLS_BACKGROUND_COLOR_DEFAULT = '#000000BB';
 
 /** default size of the icons in the Player controls area. */
 export const PLAYER_CONTROLS_ICON_SIZE_DEFAULT = '2.0rem';
+
+/** default size of the player background image. */
+export const PLAYER_BACKGROUND_IMAGE_SIZE_DEFAULT = "100% 100%";
 
 
 export const listStyle = css`